+++ src/components/icons/apps.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const AppsIcon = styled.div`
+  width: 6px;
+  height: 6px;
+  border-radius: 50%;
+  background: black;
+  border: 1px solid black;
+
+  box-shadow: 0 -15px 0 0 black, 15px -15px 0 0 black, 15px 0 0 0 black,
+    15px 15px 0 0 black, 0 15px 0 0 black, -15px 15px 0 0 black,
+    -15px 0 0 0 black, -15px -15px 0 0 black;
+`;
