+++ src/components/icons/battery.js
@@ -1,0 +1,79 @@
+import styled from "styled-components";
+
+export const BatteryEmpty = styled.div`
+  width: 25px;
+  height: 15px;
+  border: 1px solid black;
+  border-radius: 0px 2px 2px 0px;
+  position: relative;
+
+  &::after {
+    content: "";
+    position: absolute;
+    top: 4px;
+    right: -4px;
+    width: 2px;
+    height: 4px;
+    border-radius: 0px 1px 1px 0px;
+    border: 1px solid black;
+  }
+`;
+
+export const BatteryHalfEmpty = styled.div`
+  width: 25px;
+  height: 15px;
+  border: 1px solid black;
+  border-radius: 0px 2px 2px 0px;
+  position: relative;
+
+  &::before {
+    content: "";
+    position: absolute;
+    width: 15px;
+    height: 15px;
+    background: black;
+    top: 0px;
+    left: 0px;
+  }
+
+  &::after {
+    content: "";
+    position: absolute;
+    top: 4px;
+    right: -4px;
+    width: 2px;
+    height: 4px;
+    border-radius: 0px 1px 1px 0px;
+    border: 1px solid black;
+  }
+`;
+
+export const BatteryFull = styled.div`
+  width: 25px;
+  height: 15px;
+  border: 1px solid black;
+  border-radius: 0px 2px 2px 0px;
+  position: relative;
+
+  &::before {
+    content: "";
+    position: absolute;
+    width: 25px;
+    height: 15px;
+    background: black;
+    top: 0px;
+    left: 0px;
+  }
+
+  &::after {
+    content: "";
+    position: absolute;
+    top: 4px;
+    right: -4px;
+    width: 2px;
+    height: 4px;
+    border-radius: 0px 1px 1px 0px;
+    border: 1px solid black;
+    background: black;
+  }
+`;
