+++ src/components/icons/index.js
@@ -1,0 +1,47 @@
+export * from "./A";
+export * from "./apps";
+export * from "./arrow";
+export * from "./battery";
+export * from "./briefcase";
+export * from "./building";
+export * from "./circle";
+export * from "./correct";
+export * from "./creditCard";
+export * from "./cross";
+export * from "./dialouge";
+export * from "./dice";
+export * from "./dialouge";
+export * from "./divide";
+export * from "./eye";
+export * from "./file";
+export * from "./focus";
+export * from "./focus";
+export * from "./folder";
+export * from "./globe";
+export * from "./gmail";
+export * from "./image";
+export * from "./instagram";
+export * from "./laptop";
+export * from "./lines";
+export * from "./lock";
+export * from "./menu";
+export * from "./minus";
+export * from "./mobile";
+export * from "./moon";
+export * from "./multiply";
+export * from "./notAllowed";
+export * from "./openInNewTab";
+export * from "./plus";
+export * from "./profile";
+export * from "./react";
+export * from "./rectangle";
+export * from "./search";
+export * from "./square";
+export * from "./star";
+export * from "./sun";
+export * from "./triangle";
+export * from "./tripleDots";
+export * from "./usb";
+export * from "./window";
+export * from "./wrong";
+export * from "./zoom";
