+++ src/components/icons/minus.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const MinusSign = styled.div`
+  width: 7px;
+  height: 40px;
+  background: black;
+  position: relative;
+  border-radius: 2px;
+  transform: rotate(270deg);
+`
