+++ src/components/icons/plus.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const PlusSign = styled.div`
+  width: 7px;
+  height: 40px;
+  background: black;
+  position: relative;
+  border-radius: 2px;
+  
+  &::before{
+    content: '';
+    width: 7px;
+    height: 40px;
+    position: absolute;
+    top: 0px;
+    background: black;
+    transform: rotate(270deg);
+    border-radius: 2px;
+  }
+`
