+++ src/components/icons/rectangle.js
