+++ src/components/icons/search.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const SearchIcon = styled.div`
+  width: 20px;
+  height: 20px;
+  border-radius: 50%;
+  border: 3px solid black;
+  position: relative;
+
+  &:after{
+    content: '';
+    position: absolute;
+    width: 3px;
+    height: 12px;
+    background: black;
+    bottom: -9px;
+    right: -5px;
+    transform: rotate(-45deg);
+  }
+`
