+++ src/components/icons/triangle.js
