+++ src/components/index.js
@@ -1,8 +1,194 @@
-export * from "./circle";
-export * from "./lines";
-export * from "./triangle";
-export * from "./square";
-export * from "./rectangle";
-export * from "./signs";
-export * from "./alphabets";
-export * from "./boxShadowIcons";
+import {
+    Circle,
+    SemiCircle,
+    HorizontalLine,
+    VerticalLine,
+    SolidTriangle,
+    Square,
+    Reactange,
+    CrossSign,
+    Star,
+    AppsIcon,
+    SunSolid,
+    SunSolidHorizon,
+    ArrowToLeft,
+    ArrowToRight,
+    ArrowToDown,
+    ArrowToUp,
+    TripleDotsHorizontal,
+    TripleDotsVertical,
+    MenuIconHorizontal,
+    FileIcon,
+    Folder,
+    Profile,
+    BatteryEmpty,
+    BatteryHalfEmpty,
+    BatteryFull,
+    Mobile,
+    Laptop,
+    LaptopWithText,
+    Building,
+    NotAllowed,
+    CreditCard,
+    DiceOne,
+    DiceTwo,
+    DiceThree,
+    DiceFour,
+    DiceFive,
+    DiceSix,
+    OpenInNewTab,
+    WindowsIcon,
+    PlusSign,
+    MinusSign,
+    MultiplySign,
+    DivideSign,
+    Gmail,
+    Instagram,
+    Focus,
+    ReactIcon,
+    BriefCase,
+    Globe,
+    Moon,
+    Correct,
+    Wrong,
+    SearchIcon,
+    ZoomInIcon,
+    ZoomOutIcon,
+    USB,
+    Eye,
+    Image,
+    Lock,
+    LockOpen
+  } from "./icons";
+
+export const ComponentsMap = {
+    // dark icons
+    dice_one: DiceOne,
+    dice_two:DiceTwo,
+    dice_three:DiceThree,
+    dice_four:DiceFour,
+    dice_five:DiceFive,
+    dice_six:DiceSix,
+    apps: AppsIcon,
+    sun_solid:SunSolid,
+    sun_solid_horizon:SunSolidHorizon,
+    arrow_to_left:ArrowToLeft,
+    arrow_to_right:ArrowToRight,
+    arrow_to_down:ArrowToDown,
+    arrow_to_up:ArrowToUp,
+    file: FileIcon,
+    folder: Folder,
+    building: Building,
+    credit_card: CreditCard,
+    windows: WindowsIcon,
+    solid_triangle:SolidTriangle,
+    plus: PlusSign,
+    minus:MinusSign,
+    multiply: MultiplySign,
+    divide: DivideSign,
+    gmail: Gmail,
+    instagram: Instagram,
+    globe: Globe,
+    moon: Moon,
+    correct: Correct,
+    wrong: Wrong,
+    search: SearchIcon,
+    zoom_in: ZoomInIcon,
+    zoom_out: ZoomOutIcon,
+    usb: USB,
+    eye: Eye,
+    image: Image,
+    lock: Lock,
+    lock_open: LockOpen,
+
+    // light icons
+    battery_full: BatteryFull,
+    battery_half_empty:BatteryHalfEmpty,
+    battery_empty: BatteryEmpty,
+    circle: Circle,
+    semi_circle: SemiCircle,
+    horizontal_line:HorizontalLine,
+    verical_line: VerticalLine,
+    square: Square,
+    rectangle: Reactange,
+    cross: CrossSign,
+    star: Star,
+    triple_dot_horizontal: TripleDotsHorizontal,
+    triple_dot_vertical: TripleDotsVertical,
+    menu_horizontal: MenuIconHorizontal,
+    prpfile: Profile,
+    mobile: Mobile,
+    laptop: Laptop,
+    laptop_with_text: LaptopWithText,
+    open_in_new_tab: OpenInNewTab,
+    not_allowed: NotAllowed,
+    focus: Focus,
+    react: ReactIcon,
+    briefcase: BriefCase
+};
+
+export const Components = [
+    // dark icons
+    DiceOne,
+    DiceTwo,
+    DiceThree,
+    DiceFour,
+    DiceFive,
+    DiceSix,
+    AppsIcon,
+    SunSolid,
+    SunSolidHorizon,
+    ArrowToLeft,
+    ArrowToRight,
+    ArrowToDown,
+    ArrowToUp,
+    FileIcon,
+    Folder,
+    Building,
+    CreditCard,
+    WindowsIcon,
+    SolidTriangle,
+    PlusSign,
+    MinusSign,
+    MultiplySign,
+    DivideSign,
+    Gmail,
+    Instagram,
+    Globe,
+    Moon,
+    Correct,
+    Wrong,
+    SearchIcon,
+    ZoomInIcon,
+    ZoomOutIcon,
+    USB,
+    Eye,
+    Image,
+    Lock,
+    LockOpen,
+
+    // light icons
+    BatteryFull,
+    BatteryHalfEmpty,
+    BatteryEmpty,
+    Circle,
+    SemiCircle,
+    HorizontalLine,
+    VerticalLine,
+    Square,
+    Reactange,
+    CrossSign,
+    Star,
+    TripleDotsHorizontal,
+    TripleDotsVertical,
+    MenuIconHorizontal,
+    Profile,
+    Mobile,
+    Laptop,
+    LaptopWithText,
+    OpenInNewTab,
+    NotAllowed,
+    Focus,
+    ReactIcon,
+    BriefCase
+  ];
