+++ src/components/layout/layoutHelper.js
@@ -35,6 +35,7 @@ export const Header = styled.div`
 `
 export const Container = styled.div`
     display: grid;
+    min-height: calc(100vh - 120px);
     gap: 20px;
     grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
     margin: 10px 20px;
